requirements.txt
-i https://pypi.org/simple
-e .
django-debug-toolbar==1.11
django-environ==0.4.5
django-guardian==1.5.0
django==2.1.7
pdbpp==0.9.7
psycopg2==2.7.7
whitenoise==4.1.2
gunicorn==19.9.0
